Using SPARK MAX for ball intake arm

+++ src/main/include/commands/MoveIntakeArmCommand.h
@@ -11,12 +11,12 @@
 
 class MoveIntakeArmCommand : public frc::Command {
 public:
-  MoveIntakeArmCommand(int position);
+  MoveIntakeArmCommand(double position);
   void Initialize() override;
   void Execute() override;
   bool IsFinished() override;
   void End() override;
   void Interrupted() override;
 private:
-  int position;
+  double position;
 };
